Fixed reason reported for startsWith #1818
