+++ src/PSRule/Common/ExpressionContextExtensions.cs
@@ -8,14 +8,14 @@ namespace PSRule;
 
 internal static class ExpressionContextExtensions
 {
-    public static bool ExpressionTrace(this ExpressionContext context, string name, object operand, object value)
+    public static bool ExpressionTrace(this IExpressionContext context, string name, object operand, object value)
     {
         var type = context.Kind == Definitions.ResourceKind.Rule ? 'R' : 'S';
         context.Debug(PSRuleResources.LanguageExpressionTraceP3, type, name, operand, value);
         return true;
     }
 
-    public static bool ExpressionTrace(this ExpressionContext context, string name, object value)
+    public static bool ExpressionTrace(this IExpressionContext context, string name, object value)
     {
         var type = context.Kind == Definitions.ResourceKind.Rule ? 'R' : 'S';
         context.Debug(PSRuleResources.LanguageExpressionTraceP2, type, name, value);
