+++ src/PSRule/Definitions/Expressions/FunctionBuilder.cs
@@ -1,19 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Diagnostics;
-
 namespace PSRule.Definitions.Expressions;
 
 internal delegate object ExpressionFnOuter(IExpressionContext context);
 internal delegate object ExpressionFn(IExpressionContext context, object[] args);
 
 internal delegate ExpressionFnOuter ExpressionBuilderFn(IExpressionContext context, LanguageExpression.PropertyBag properties);
-
-internal abstract class FunctionReader
-{
-    public abstract bool TryProperty(out string propertyName);
-}
 
 internal sealed class FunctionBuilder
 {
@@ -58,61 +51,4 @@ internal sealed class FunctionBuilder
         }
         return false;
     }
-}
-
-internal sealed class FunctionFactory
-{
-    private readonly Dictionary<string, IFunctionDescriptor> _Descriptors;
-
-    public FunctionFactory()
-    {
-        _Descriptors = new Dictionary<string, IFunctionDescriptor>(StringComparer.OrdinalIgnoreCase);
-        foreach (var d in Functions.Builtin)
-            With(d);
-    }
-
-    public bool TryDescriptor(string name, out IFunctionDescriptor descriptor)
-    {
-        return _Descriptors.TryGetValue(name, out descriptor);
-    }
-
-    public void With(IFunctionDescriptor descriptor)
-    {
-        _Descriptors.Add(descriptor.Name, descriptor);
-    }
-}
-
-/// <summary>
-/// A structure describing a specific function.
-/// </summary>
-[DebuggerDisplay("Function: {Name}")]
-internal sealed class FunctionDescriptor : IFunctionDescriptor
-{
-    public FunctionDescriptor(string name, ExpressionBuilderFn fn)
-    {
-        Name = name;
-        Fn = fn;
-    }
-
-    /// <inheritdoc/>
-    public string Name { get; }
-
-    /// <inheritdoc/>
-    public ExpressionBuilderFn Fn { get; }
-}
-
-/// <summary>
-/// A structure describing a specific function.
-/// </summary>
-internal interface IFunctionDescriptor
-{
-    /// <summary>
-    /// The name of the function.
-    /// </summary>
-    string Name { get; }
-
-    /// <summary>
-    /// The function delegate.
-    /// </summary>
-    ExpressionBuilderFn Fn { get; }
 }
