+++ src/PSRule/Definitions/Expressions/FunctionReader.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions.Expressions;
+
+internal abstract class FunctionReader
+{
+    public abstract bool TryProperty(out string propertyName);
+}
