+++ src/PSRule/Definitions/Expressions/LanguageExpressionType.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions.Expressions;
+
+internal enum LanguageExpressionType
+{
+    Operator = 1,
+
+    Condition = 2,
+
+    Function = 3
+}
