+++ tests/PSRule.Tests/Usings.cs
@@ -3,3 +3,4 @@
 
 global using Xunit;
 global using Assert = Xunit.Assert;
+global using Moq;
